data_collection/requirements.txt
confluent-kafka[avro,json,protobuf]
requests
six
pandas
pymongo[srv]==4.1.1
jsonschema==3.2.0
dnspython
python-dotenv
-e .